refactor: split python worker support helpers
