+++ src/lib/doc-runtime/python-worker-queue.ts
@@ -1,0 +1,14 @@
+export function createSerialRequestQueue<Request>(
+  handle: (request: Request) => Promise<void>
+): { enqueue: (request: Request) => void } {
+  let tail = Promise.resolve();
+
+  return {
+    enqueue(request) {
+      tail = tail
+        .catch(() => undefined)
+        .then(() => handle(request))
+        .catch(() => undefined);
+    }
+  };
+}
